requirements.txt
pytest==6.2.2
pytest-runner==5.3.0
numpy==1.20.1
Pillow==8.1.0